Move all MalSimulator related classes into one file (mal_simulator.py)
